refactor(GiniCaptureSDK): refactor camera viewcontroller
PIA-2918
